refactor(Button Systems): Refactor the way button systems are created
Since I created an abstract for the buttons init systems, I updated those button systems
