+++ R-TypeGame/Systems/Menus/Buttons/IpTextBox/IpTextBoxInitSystem.hpp
@@ -1,0 +1,46 @@
+/*
+** EPITECH PROJECT, 2024
+** R-Type
+** File description:
+** IpTextBoxInitSystem.hpp
+*/
+
+#pragma once
+
+#include "ISystem.hpp"
+#include "AButtonInitSystem.hpp"
+
+#define PATH_JSON "GameData/Scenes/Menus/Buttons/ipTextBoxSystem.json"
+
+class IpTextBoxInitSystem : public AButtonInitSystem, public ASystem {
+
+    public:
+
+        /**
+         * @brief Construct a new IpTextBoxInitSystem object
+         *
+         */
+        IpTextBoxInitSystem();
+
+        /**
+         * @brief Destroy the IpTextBoxInitSystem object
+         *
+         */
+        ~IpTextBoxInitSystem() = default;
+
+        /**
+         * @brief Get the Function object.
+         *
+         * @return std::function<void(ECS::Registry& reg, int idxPacketEntities)>
+         */
+        std::function<void(ECS::Registry &reg, int idxPacketEntities)> getFunction()
+        {
+            return [this](ECS::Registry& reg, int idxPacketEntities) {
+                _initButton(reg, idxPacketEntities);
+            };
+        }
+
+    private:
+
+        void _initButton(ECS::Registry& reg, int idxPacketEntities);
+};
